src/domain/notification/application/use-cases/send-notification.ts
import { right } from '@/core/either'
import { SendNotificationUseCaseRequestDTO } from './dto/send-notification-request'
import { SendNotificationUseCaseResponseDTO } from './dto/send-notification-response'
import { Notification } from '../../enterprise/entities/notification'
import { UniqueEntityID } from '@/core/entities/unique-entity-id'
import { NotificationsRepository } from '../repositories/notifications-repository'

export class SendNotificationUseCase {
  constructor(
    private readonly notificationsRepository: NotificationsRepository,
  ) {}

  async execute({
    recipientId,
    title,
    content,
  }: SendNotificationUseCaseRequestDTO): Promise<SendNotificationUseCaseResponseDTO> {
    const notification = Notification.create({
      recipientId: new UniqueEntityID(recipientId),
      content,
      title,
    })

    await this.notificationsRepository.create(notification)

    return right({
      notification,
    })
  }
}
